install.sh skips symlinking unwanted files.

+++ install.sh
@@ -2,18 +2,35 @@
 
 backup_directory="$HOME/.dotfiles.backup"
 dotfiles_directory="$HOME/.dotfiles"
+files_to_skip=("install.sh" ".git")
 
 create_backup_dir() {
     mkdir -p "$backup_directory"
+}
+
+should_skip_this_file() {
+    local name="$1"
+
+    for skip in "${files_to_skip[@]}"; do
+        if [[ "$name" == "$skip" ]]; then
+            return 0    # yes
+        fi
+    done
+
+    return 1    # no
 }
 
 move_identical_files_or_dirs_to_backup_dir() {
     shopt -s nullglob dotglob
     for item in "$dotfiles_directory/"*; do
 	local file_or_dir_name=$(basename "$item")
+
 	if [[ -e "$HOME/$file_or_dir_name" ]]; then
-	    mv "$HOME/$file_or_dir_name" "$backup_directory"
+    		if ! should_skip_this_file "$file_or_dir_name"; then
+        		mv "$HOME/$file_or_dir_name" "$backup_directory"
+    		fi
 	fi
+
     done
     shopt -u nullglob dotglob
 }
@@ -23,6 +40,10 @@ create_symlinks_to_dotfiles() {
     for item in "$dotfiles_directory/"*; do
         local name
         name=$(basename "$item")
+
+	if should_skip_this_file "$name"; then
+		continue
+	fi	
 
         if [[ $1 == "overwrite" ]]; then
             ln -sfn "$dotfiles_directory/$name" "$HOME/$name"
